yoloV8_train/requirements.txt
numpy
opencv-python-headless
Pillow
argparse
